requirements.txt
python == 3.13.5
langchain == 0.3.26
google-generativeai == 0.8.5
chromadb == 1.0.13
faiss-cpu == 1.11.0
unstructured == 0.17.2
pypdf == 5.6.1
fastapi == 0.115.13
uvicorn == 0.34.3
tiktoken == 0.9.0
python-dotenv == 1.1.1
PyMuPDF == 1.26.1
langchain-google-genai == 2.0.10
langchain-community == 0.3.26
streamlit == 1.46.0
langchain_chroma == 0.2.4